Audio/VR Fixes, Added Controller support:
- Fixed audio files not loading on Linux
- Fixed VR not stopping if an error occured on initialization
- Moved/added environnement position to VRUtils VRUtils::getEnvPosition
- Fixed VR not respecting envPos (It's now inclued in VRUtils::getPosition, VRUtils::getViewMatrix and VRUtils::getHmdPose)
- Fixed inverted audio (left/right)
- Fixed bad forward direction (inverted)
- Added XBox controller support
- Removed DepthBuffer from frames sent to HMD
- Added somes parameters to Configuration

+++ src/slaynash/opengl/render2d/GUIManager.java
@@ -312,8 +312,8 @@ public class GUIManager {
 		float aspectRatio = DisplayManager.getWidth()/((float)DisplayManager.getHeight());
 		float imgAspectRatio = imageWidth/imageHeight;
 		
-		float maxX = (imageWidth/(float)(background.getTexture().getTextureWidth()));
-		float maxY = (imageHeight/(float)(background.getTexture().getTextureHeight()));
+		float maxX = (imageWidth/(background.getTexture().getTextureWidth()));
+		float maxY = (imageHeight/(background.getTexture().getTextureHeight()));
 		float centerx = maxX*0.5f;
 		float centery = maxY*0.5f;
