+++ src/slaynash/opengl/render2d/text2d/fontMeshCreator/TextMeshCreator.java
@@ -30,7 +30,7 @@ public class TextMeshCreator {
 		Line currentLine = new Line(metaData.getSpaceWidth(), text.getFontSize(), text.getMaxLineSize());
 		Word currentWord = new Word(text.getFontSize());
 		for (char c : chars) {
-			int ascii = (int) c;
+			int ascii = c;
 			if (ascii == SPACE_ASCII) {
 				boolean added = currentLine.attemptToAddWord(currentWord);
 				if (!added) {
